Cleanup
Scripts with privatised parameters are now separate scripts to avoid breaking of other actually networked scripts that rely on each other.
New prefabs:
_Mirror
_Settings
