+++ Control Box Example/Udon/DrawerElement.cs
@@ -46,14 +46,14 @@ public class DrawerElement : UdonSharpBehaviour
         }
     }
 
-    public void _SendMaxEvent()
+    public void SendMaxEvent()
     {
         maxevent.color = Color.green;
         eventCooldown = true;
         audioSource.PlayOneShot(endSound);
     }
 
-    public void _SendMinEvent()
+    public void SendMinEvent()
     {
         minevent.color = Color.green;
         eventCooldown = true;
