+++ Control Box Example/Udon/SliderElement.cs
@@ -64,7 +64,7 @@ public class SliderElement : UdonSharpBehaviour
         }
     }
 
-    public void _SendMaxEvent()
+    public void SendMaxEvent()
     {
         maxevent.color = Color.green;
         eventCooldown = true;
@@ -72,7 +72,7 @@ public class SliderElement : UdonSharpBehaviour
         audioSource.PlayOneShot(clickEnd);
     }
 
-    public void _SendMinEvent()
+    public void SendMinEvent()
     {
         minevent.color = Color.green;
         eventCooldown = true;
