+++ Mirror/_MirrorToggle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 68d25f451de8b49429dfc554c58a2fb3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
