+++ Prefabs/_Settings.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6c64c601ab6c34a4bab55dcfa30e20ac
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
