+++ Udon/LightIntensity.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 45995d34091d6524faf4d8b12795965a
+guid: 7f70c78d30359ca419dce7b88a58e948
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
