+++ Udon/LightIntensity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 133fbf13e88f4ee47986a2183df145d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
