+++ Udon/_PushButton.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6796055332fefaf449e6e10bd0fad484
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
